+++ tests/app/Http/Controllers/HandoffControllerTest.php
@@ -12,4 +12,10 @@ class HandoffControllerTest extends BaseApiTestCase
         $this->makeUnauthenticatedApiRequest(self::METHOD_GET, 'handoffs/dependency')
             ->assertJson($this->app->make(HandoffService::class)->getHandoffsV2Dependency());
     }
+
+    public function testItGetsHandoffV2Dependency()
+    {
+        $this->makeUnauthenticatedApiRequest(self::METHOD_GET, 'handoffs/dependency')
+            ->assertJson($this->app->make(HandoffService::class)->getHandoffsV2Dependency());
+    }
 }