+++ tests/util/Traits/WithNetworkController.php
@@ -2,17 +2,19 @@
 
 namespace util\Traits;
 
+use App\Models\Controller\ControllerPosition;
 use App\Models\Vatsim\NetworkControllerPosition;
 
 trait WithNetworkController
 {
-    public function setNetworkController(int $cid, int $controllerPositionId = 1)
+    public function setNetworkController(int $cid = self::ACTIVE_USER_CID, int $controllerPositionId = 3)
     {
+        $position = ControllerPosition::findOrFail($controllerPositionId);
         NetworkControllerPosition::upsert(
             [
                 'cid' => $cid,
-                'callsign' => 'FOO_TWR',
-                'frequency' => 123.456,
+                'callsign' => $position->callsign,
+                'frequency' => $position->frequency,
                 'controller_position_id' => $controllerPositionId,
             ],
             ['cid']